# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M15-SchoolTech API — Plateforme de gestion scolaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/contacts:
    get:
      operationId: getContactsDisponibles
      tags: [messages]
      summary: Contacts disponibles selon rôle
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste des contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsResponse"
  /messages/non-lus/count:
    get:
      operationId: getNbMessagesNonLus
      tags: [messages]
      summary: Nombre de messages non lus
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Compteur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageCountResponse"
  /messages/reception:
    get:
      operationId: getBoiteReception
      tags: [messages]
      summary: Boîte de réception
      security: [{ bearerAuth: [] }]
      parameters:
        - name: lu
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Messages reçus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesListeResponse"
  /messages/envoi:
    get:
      operationId: getBoiteEnvoi
      tags: [messages]
      summary: Boîte d'envoi
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Messages envoyés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesListeResponse"
  /messages/envoyer:
    post:
      operationId: envoyerMessage
      tags: [messages]
      summary: Envoyer un nouveau message
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvoyerMessageInput"
      responses:
        "201":
          description: Message envoyé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItemResponse"
  /messages/{id}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Lire un message (marque comme lu)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Message avec réponses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetailResponse"
  /messages/{id}/repondre:
    post:
      operationId: repondreMessage
      tags: [messages]
      summary: Répondre à un message
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepondreMessageInput"
      responses:
        "201":
          description: Réponse envoyée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItemResponse"
  /messages/{id}/archiver:
    put:
      operationId: archiverMessage
      tags: [messages]
      summary: Archiver un message
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Message archivé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Rendez-vous ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    MessageItem:
      type: object
      properties:
        id: { type: string }
        etablissement_id: { type: string }
        expediteur_id: { type: string }
        expediteur_nom: { type: string }
        expediteur_prenoms: { type: string }
        expediteur_role: { type: string }
        destinataire_id: { type: string }
        destinataire_nom: { type: string }
        destinataire_prenoms: { type: string }
        sujet: { type: string }
        contenu: { type: string }
        lu: { type: boolean }
        date_lecture: { type: string, nullable: true }
        piece_jointe_url: { type: string, nullable: true }
        piece_jointe_nom: { type: string, nullable: true }
        parent_message_id: { type: string, nullable: true }
        archive_expediteur: { type: boolean }
        archive_destinataire: { type: boolean }
        created_at: { type: string }
    MessagesListeResponse:
      type: object
      required: [messages, total]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageItem"
        total: { type: integer }
    MessageItemResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/MessageItem"
    MessageDetailResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/MessageItem"
        reponses:
          type: array
          items:
            $ref: "#/components/schemas/MessageItem"
    EnvoyerMessageInput:
      type: object
      required: [destinataire_id, sujet, contenu]
      properties:
        destinataire_id: { type: string }
        sujet: { type: string }
        contenu: { type: string }
        piece_jointe_url: { type: string }
        piece_jointe_nom: { type: string }
        parent_message_id: { type: string }
    RepondreMessageInput:
      type: object
      required: [contenu]
      properties:
        contenu: { type: string }
        piece_jointe_url: { type: string }
        piece_jointe_nom: { type: string }
    ContactItem:
      type: object
      properties:
        id: { type: string }
        nom: { type: string }
        prenoms: { type: string }
        role: { type: string }
        photo_url: { type: string, nullable: true }
    ContactsResponse:
      type: object
      required: [contacts]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactItem"
    MessageCountResponse:
      type: object
      required: [count]
      properties:
        count: { type: integer }

    # ── Rendez-vous ────────────────────────────────────────────
...